Cygwin: devdoc: Mention the extremely useful small_printf() function
This function came in real handy many, many times over the year whenever
I needed to debug any issues with the MSYS2/Cygwin runtime, first
instance was while debugging an issue that strace 'fixed'.

However, this function was not mentioned anywhere I looked, so it took
me a good while to find out about it. Let's improve on that situation by
mentioning it explicitly in the documentation about debugging the
runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>