libgloss: mips: g++ exception handling failure on o32 elf targets
The compiler driver positions the linker script at the end of the linker
command-line, after crtend.o. As a result, any INPUT objects and archive
GROUPs introduced by the linker script end up after crtend.o.  This messes up
the end-of-frame marker provided by crtend.o

This has always been a problem, but a recent binutils update to clean-up
redundant NULL markers in .eh_frame exposes it as a execution failure in
exception-handling tests. This patch re-orders .eh_frame in all linker
scripts so that the one from crtend.o is placed last.

An alternative approach would be to fix the compiler driver to put
the linker script before the ENDFILE spec.

Signed-off-by: Jovan Dmitrović <jovan.dmitrovic@htecgroup.com>