Cygwin: AF_UNIX: add methods to check for unread data
When reading from a stream socket, it's possible that the caller
requested fewer bytes that are available in the packet.  In this case
there is user data left in the pipe without a packet header.

Add a member '_unread' to the af_unix_shmem_t class to keep track of
this, and add corresponding methods 'get_unread' and 'set_unread' to
af_unix_shmem_t and fhandler_socket_unix.