Cygwin: AF_UNIX: fix MAX_AF_PKT_LEN
Define it to be UINT16_MAX, since the pckt_len member of
af_unix_pkt_hdr_t is of type uint16_t.

Previously it was 65536 (== UINT16_MAX + 1), which is 0 when cast to
uint16_t.  So an attempt to write a packet of size MAX_AF_PKT_LEN
actually wrote 0 bytes.