Cygwin: AF_UNIX: SCM_RIGHTS: support console descriptors
This requires modifying fhandler_console::dup so that it duplicates
handles when it is called with non-zero src_pid on an fhandler that
doesn't yet have an archetype.

As in the recent change to fhandler_pty_slave::dup, dtable::dup_worker
will then create an archetype for the new fhandler after
fhandler_console::dup returns.

Untested.