Cygwin: AF_UNIX: update tests
The previous test, send_tty, didn't really exercise the new
fhandler_pty_slave::dup code, since the descriptor that was sent was
for a pty slave already open in the subprocess.  So it already had an
archetype, and no handles were duplicated.

Replace it by a new test, send_pty_slave, that does exercise the new
code (successfully).