Cygwin: poll(2): extend the socket code to all sockets
poll has code that checks whether a socket descriptor has a failed
connect or whether shutdown has been called.  This previously worked
only for wsock sockets.  Tweak the code so that it applies to all
sockets.

To help with this, add a virtual saw_shutdown_read method to the
fhandler_socket class, and define a version of it for AF_UNIX sockets.
A version for wsock sockets already existed.

Untested.