Cygwin: AF_UNIX: sendmsg: wait for a pipe instance in the datagram case
When trying to connect to the pipe created by the remote socket,
open_pipe might fail if there are no pipe instances available.  This
can happen if there has been a previous connection to the same pipe.
The server end of the pipe won't be available again until it
disconnects the previous connection.  (This is done in recvmsg.)

Introduce and use a new method fhandler_socket_unix::wait_open_pipe
that waits (in the blocking case) for an instance to become available.