Cygwin: AF_UNIX: select: fix some errors
Try to make the implementation compatible with Stevens, UNIX Network
Programming.

select.cc:peek_socket_unix now calls
fhandler_socket_unix::binding_state and
fhandler_socket_unix::so_error, so these need to be made public.

One glaring remaining problem is that our current implementation of
datagram sockets makes it impossible to check whether a write would
block.  The problem is that we have no send buffer until we connect to
a pipe.

Also change the use of except_on_write so that it is set only if the
connect state is connect_pending when select is called.  Use this to
check whether an async connect has completed, which implies ready for
writing according to Stevens.