Cygwin: AF_UNIX: adapt symlink and reparse point to mqueues
create_symlink, create_reparse_point, and create_file now take a char
*mqueue_name argument instead of a PUNICODE_STRING.  For
create_symlink, convert the string to a UNICODE_STRING and make that
the target of the symlink.  For create_reparse_point, store the string
in the reparse point as is.

Make other provisional changes so that compilation will succeed.  This
includes adding a methode connect_mqueue that does nothing for now.