Cygwin: AF_UNIX: keep a reference count of open descriptors
Add a data member _ndesc to the af_unix_shmem_t class, along with
methods to increment and decrement it.

Increment it during socket/fork/exec/dup, and decrement it during
close.  When the last descriptor is closed, call shutdown (SHUT_RDWR)
and unlink the socket's message queue.

The shutdown call will make the socket's peer see EPIPE if it tries to
write.