Cygwin: console: Fix tcsetattr which was broken sinse cygwin 3.5.5
After the commit 84d77e5918e1 ("Cygwin: console: Disable
cons_master_thread in win32-input-mode") tcsetattr() no longer takes
effect for the console. This occurs because set_input_mode() and
set_output_mode() are now called in bg_check() only when the current
mode is not tty::cygwin. Before this commit, these functions were
always invoked in bg_check(), regardless of whether the current mode
was already tty::cygwin. As a result, the console mode was updated
every time read() or write() was called.

With this patch, set_input_mode() and set_output_mode() are called
in tcsetattr() to make it take effect.

Fixes: 84d77e5918e1 ("Cygwin: console: Disable cons_master_thread in win32-input-mode")
Signed-off-by: Takashi Yano <takashi.yano@nifty.ne.jp>