Cygwin: Only return true from try_to_debug() if we launched a JIT debugger
This fixes constantly replaying the exception if we have a segfault
while a debugger is already attached, e.g. stracing a segv, see:

https://cygwin.com/pipermail/cygwin/2025-May/258144.html

Future work:

The 'debugging' static in exception::handle(), which makes us replay the
exception the next half a million times it's hit seems like cruft, maybe
we should look at if it's possible to remove that?

Fixes: 91457377d6c9 ("Cygwin: Make 'ulimit -c' control writing a coredump")
Reported-by: Christian Franke <Christian.Franke@t-online.de>
Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>