Cygwin: doc: Install FAQ as well
We don't currrently deploy the CSS file to the website, because the FAQ
doesn't get installed, so instead we copy all the doc files from the
build directory.

That omits docbook.css, which is installed from directly from the source
directory. (We just happen to have a similar file lying around on the
webserver).

Instead, just install the FAQ.

There will be a separate change to the cygwin packaging to avoid
including the FAQ in the cygwin-doc package. (I guess we'd rather people
go online for that, to ensure they have the latest version?)

Also, add matching uninstall rules.