Cygwin: add fast-path for posix_spawn(p)
Currently just file actions are supported in the fast path, open/dup2
with a target fd of stdin/out/err, close of any fd, and chdir and
fchdir.  These were chosen as the least-common-denominator of
functionality supported for any child process, as they are allowed to be
specified via CreateProcess.  They also happen to be the most common
file operations, performed by the likes of GNU make, ninja, LLVM, and rust.

For other attributes or actions, fall back to the newlib implementation
of posix_spawn (which uses fork/exec to set everything up in the child
process).

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>