libgloss: Fix invalid prototype
The __inln prototype was defined as:

extern char * __inln ();

In old versions of C, this meant "this function takes an unspecified
number of parameters". This seems to have changed, as GCC 14 now
considers that the function took zero parameters.

As the code called it with one parameter, GCC would then issue an error
saying that the incorrect number of parameters was passed.

Address this issue by changing the prototype of the __inln function.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>