Cygwin: make pthread initializer macros constinit compliant
In order to avoid a restriction on any reinterpret_cast-like behavior in
constinit expressions, use assembly and the linker to define symbols
with the not-valid-address addresses.

Due to being built with -mcmodel=small, this linker trick does not
always work in the Cygwin DLL, specifically in the cases where they are
used in comparisons (==).  As a result, leave the case where the
symbols would have had to have been classes rather than structs as casts
of integers.  If Cygwin ever needs these to be constinit compliant, it
may need to move to the medium code model or add some other workaround.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>