Cygwin: testsuite: test posix_file_actions_add(f)chdir
Also test their interaction with addopen, as opens added subsequent to a
chdir need to be relative to that new cwd.

In order for the tests to compile on Linux, define O_SEARCH to O_PATH if
O_SEARCH is not defined, and use the *chdir_np names instead of the
now-standardized *chdir.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>