Cygwin: doc: Install miscellaneous website files
This gives us a full set of the built documentation files in the install
directory, which can then be appropriately filtered for packaging or
deploying to website, as required.

Again, there will be a separate change to the cygwin packaging to avoid
including theses files in the cygwin-doc package.

Also, add matching uninstall rules.