mbrtowc: fix handling invalid UTF-8 4 byte sequences if wchar_t == UTF-16
When commit 28186e81d947 split _mbtowc_r into per-codeset functions, the
code generating wchar_t from UTF-8 input was slightly rearranged.
Unfortunately the new code introduced a bug:

On systems with wchar_t being UTF-16, 4 byte sequences have to be
converted to surrogate pairs.  The low surrogate pair can be fully
created from the first 3 bytes of the sequence.  However, the surrogates
should only be created if it's clear that the 4th byte is valid, and the
entire 4 byte string represents a valid UTF-8 sequence.

The code change in 28186e81d947 neglected just that: In contrast to the
original code, it now created the low surrogate after having read the
first 3 bytes of the sequence, without checking validity of the 4th byte.

This patch moves the test sequence to check the 4th byte in front of the
code generating the low surrogate.  Make sure to return the value 3 (3
bytes digested) rather than the content of the local variable i, which
is already set to 4 at this point.

Reported-by: Christian Franke <Christian.Franke@t-online.de>
Addresses: https://cygwin.com/pipermail/cygwin/2025-June/258358.html
Fixes: 28186e81d947 ("* libc/ctype/iswalpha.c: Handle all wchar_t as unicode on _MB_CAPABLE systems.")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>