Cygwin: drop ASCII NUL from invalid DOS filename char transposition table
The mechanism to convert characters invalid in DOS/Windows filenames
to the Unicode private use area starting at 0xf000 accidentally also
converts ASCII NUL to the private use area.  That usually doesn't happen,
but what if the filename contains the 0xf000 character and we want to
convert it back to multibyte?  In this case the 0xf000 becomes the ASCII
NUL again, and the resulting string is cut short and does not match with
the real filename.  This leads to all sorts of problems like the
inability to unlink the file.

Fix this problem by not marking ASCII NUL as transposable character.

This problem has been introduced by adding code to support abstract
sockets which may contain a NUL char as part of the socket name.  By
reverting the ASCII NUL in the transposition tables,
transform_chars_af_unix has to handle ASCII NUL explicitely now.

Reported-by: Christian Franke <Christian.Franke@t-online.de>
Addresses: https://cygwin.com/pipermail/cygwin/2025-June/258367.html
Fixes: 7d260cfac42d ("Cygwin: add transform_chars_af_unix helper")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>