Cygwin: add pgroup support to posix_spawn fast path
Tweak proc_subproc PROC_ADD_CHILD to only initialize vchild->pgid if
it's not already set.

The error checking of setpgid is lacking with respect to the POSIX
standard, but this code replicates what setpgid does.

This attribute is used by ninja, so is worth adding to the fast path.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>