cygwin: faq-programming-6.21 ready-made download commands
Running setup-x86_64.exe and performing an individual search for each package,
is a pain in the neck. And the user running "setup-x86_64.exe -q -P packageName"
may have a typo and the package would not be installed.

So by copying and pasting these commands into the terminal setup-x86_64.exe runs
a preliminary search, so when they reach the "Select Packages" window these packages
will be amongst the packages ready to be downloaded. Additionally sorting them makes it
easier to go down the list, and make sure all packages are listed.

Putting separate commands allows the user to pick which packages to download in the
case they want to run config, where those packages are not needed i.e. --without-cross-bootstrap.

Signed-off-by: John Haugabook <johnhaugabook@gmail.com>