Cygwin: console: Set console mode only if std{in,out,err} is console
Currently, when cygwin app is launched, the console input mode is
set to tty::cygwin, even if the stdin is not a console. However,
it is not necessary because the cygwin app does not use stdin.
This also applies to stdout and stderr.

With this patch, the console mode is set only when std{in,out,err}
is a console for the cygwin app for better coexistence with non-
cygwin apps.

Signed-off-by: Takashi Yano <takashi.yano@nifty.ne.jp>